Change chat_id, message_id, update_id, user_id from int to long

+++ src/main/java/com/payneteasy/telegram/bot/client/messages/ChatActionRequest.java
@@ -17,7 +17,7 @@ public class ChatActionRequest {
      * Unique identifier for the target chat
      */
     @SerializedName("chat_id")
-    private final Integer chatId;
+    private final Long chatId;
 
 //    /**
 //     * username of the target channel (in the format @channelusername)
